chore: move shared function to the cloud package
